% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{add_body}
\alias{add_body}
\title{Add a table body (a df) to the tab.}
\usage{
add_body(tab, df, left_header_colnames = NULL, row_style_names = NULL,
  left_header_style_names = NULL, col_style_names = NULL)
}
\arguments{
\item{df}{For a single top_header row, a character vector.  For multiple top_header rows, a list of character vectors.}

\item{left_header_colnames}{The names of the columns in the df which are left headers, as opposed to the main body
of the table}
}
\description{
Add a table body (a df) to the tab.
}
