% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{add_title}
\alias{add_title}
\title{Add titles to the tab.  Title text is provided as a character vector, with each element being a row of the title}
\usage{
add_title(tab, title_text, title_style_names = NULL)
}
\arguments{
\item{tab}{The core tab object}

\item{title_text}{A character vector.  Each element is a row of the title}

\item{title_style_names}{A character vector.  Each elemment is a style_name}
}
\description{
Add titles to the tab.  Title text is provided as a character vector, with each element being a row of the title
}
