% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{auto_merge_footer_cells}
\alias{auto_merge_footer_cells}
\title{Take a tab and merge the footer rows}
\usage{
auto_merge_footer_cells(tab)
}
\arguments{
\item{tab}{The core tab object}
}
\description{
Take a tab and merge the footer rows
}
\examples{
crosstab <- read.csv(system.file("extdata", "example_crosstab.csv", package="xltabr"))
tab <- initialise()
tab <- add_body(tab, crosstab)
footer_text <- c("Footer contents 1", "Footer contents 2")
footer_style_names <- c("subtitle", "subtitle")
tab <- add_footer(tab, footer_text, footer_style_names)
tab <- auto_merge_footer_cells(tab)
}
