% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{auto_merge_title_cells}
\alias{auto_merge_title_cells}
\title{Take a tab and merge the title rows}
\usage{
auto_merge_title_cells(tab)
}
\arguments{
\item{tab}{The core tab object}
}
\description{
Take a tab and merge the title rows
}
\examples{
crosstab <- read.csv(system.file("extdata", "example_crosstab.csv", package="xltabr"))
tab <- initialise()
tab <- add_body(tab, crosstab)
title_text <- c("Main title on first row", "subtitle on second row")
title_style_names <- c("title", "subtitle")
tab <- add_title(tab, title_text, title_style_names)
tab <- auto_merge_title_cells(tab)
}
