% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_style.R
\name{auto_style_number_formatting}
\alias{auto_style_number_formatting}
\title{Use the data type of the columns to choose an automatic Excel format
Number formats are only applied to body cells}
\usage{
auto_style_number_formatting(tab, overrides = list())
}
\arguments{
\item{tab}{a table object}

\item{overrides}{a list containing any manual overrides where the user wants to provide their own style name}
}
\description{
Note this reads currency styling from the styles defined \href{https://github.com/moj-analytical-services/xltabr/blob/dev/inst/extdata/number_formats.xlsx?raw=true}{here}
}
\examples{
auto_style_number_formatting(overrides = list("colname1" = "currency1"))
}
