% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{body_get_bottom_wb_row}
\alias{body_get_bottom_wb_row}
\title{Compute the bottom (lowest) row of the workbook occupied by the body
If the body does not exist, returns the last row of the previous element (the top header)}
\usage{
body_get_bottom_wb_row(tab)
}
\arguments{
\item{tab}{The core tab object}
}
\description{
Compute the bottom (lowest) row of the workbook occupied by the body
If the body does not exist, returns the last row of the previous element (the top header)
}
