% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{body_get_cell_styles_table}
\alias{body_get_cell_styles_table}
\title{Create table |row|col|style name| containing the styles names}
\usage{
body_get_cell_styles_table(tab)
}
\description{
Create table |row|col|style name| containing the styles names
}
