% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{body_get_rightmost_wb_col}
\alias{body_get_rightmost_wb_col}
\title{Compute the rightmost column of the workbook which is occupied by the body}
\usage{
body_get_rightmost_wb_col(tab)
}
\arguments{
\item{tab}{The core tab object}
}
\description{
Compute the rightmost column of the workbook which is occupied by the body
}
