% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{body_get_wb_cols}
\alias{body_get_wb_cols}
\title{Compute the columns of the workbook which are occupied by the body}
\usage{
body_get_wb_cols(tab)
}
\arguments{
\item{tab}{The core tab object}
}
\description{
Compute the columns of the workbook which are occupied by the body
}
