% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{body_get_wb_left_header_cols}
\alias{body_get_wb_left_header_cols}
\title{Compute the columns of the workbook which are occupied by the left header columns (which are a subset of the body columns)}
\usage{
body_get_wb_left_header_cols(tab)
}
\arguments{
\item{tab}{The core tab object}
}
\description{
Compute the columns of the workbook which are occupied by the left header columns (which are a subset of the body columns)
}
