% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{body_get_wb_rows}
\alias{body_get_wb_rows}
\title{Compute the rows of the workbook which are occupied by the body}
\usage{
body_get_wb_rows(tab)
}
\arguments{
\item{tab}{The core tab object}
}
\description{
Compute the rows of the workbook which are occupied by the body
}
