% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{body_initialise}
\alias{body_initialise}
\title{Create all the required properties for the top headers on the tab object}
\usage{
body_initialise(tab)
}
\arguments{
\item{tab}{The core tab object}
}
\description{
Create all the required properties for the top headers on the tab object
}
