% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{body_write_rows}
\alias{body_write_rows}
\title{Write all the body data to the workbook (but do not write style information)}
\usage{
body_write_rows(tab)
}
\arguments{
\item{tab}{The core tab object}
}
\description{
Write all the body data to the workbook (but do not write style information)
}
