% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined.R
\name{combine_all_styles}
\alias{combine_all_styles}
\title{Get all the styles currently in use by the workbook}
\usage{
combine_all_styles(tab)
}
\description{
Get all the styles currently in use by the workbook
}
