% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined.R
\name{extent_initialise}
\alias{extent_initialise}
\title{Create all the required properties of the extent}
\usage{
extent_initialise(tab, topleft_row, topleft_col)
}
\description{
Create all the required properties of the extent
}
