% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer.R
\name{footer_get_bottom_wb_row}
\alias{footer_get_bottom_wb_row}
\title{Get the bottom row of the footer in the wb}
\usage{
footer_get_bottom_wb_row(tab)
}
\description{
Get the bottom row of the footer in the wb
}
