% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer.R
\name{footer_get_cell_styles_table}
\alias{footer_get_cell_styles_table}
\title{Create table |row|col|style name| containing the styles names}
\usage{
footer_get_cell_styles_table(tab)
}
\description{
Create table |row|col|style name| containing the styles names
}
