% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer.R
\name{footer_get_rightmost_wb_col}
\alias{footer_get_rightmost_wb_col}
\title{Get the rightmost column of the footers in the wb}
\usage{
footer_get_rightmost_wb_col(tab)
}
\description{
Get the rightmost column of the footers in the wb
}
