% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer.R
\name{footer_initialise}
\alias{footer_initialise}
\title{Create all the required properties for the footer on the tab object}
\usage{
footer_initialise(tab)
}
\description{
Create all the required properties for the footer on the tab object
}
