% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set_global_opts.R
\name{get_cell_format_path}
\alias{get_cell_format_path}
\title{Get cell formats path}
\usage{
get_cell_format_path()
}
\description{
Returns path to cell format definitions which is set in global options. To change this path see set_cell_format_path.
Default cell format used by package is \href{https://github.com/moj-analytical-services/xltabr/blob/dev/inst/extdata/style_to_excel_number_format.csv?raw=true}{here}
}
