% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set_global_opts.R
\name{get_num_format_path}
\alias{get_num_format_path}
\title{Get number formats path}
\usage{
get_num_format_path()
}
\description{
Returns path to number format definitions which is set in global options. To change this path see get_num_format_path.
Default number formats used by package is \href{https://github.com/moj-analytical-services/xltabr/blob/master/inst/extdata/number_format_defaults.csv?raw=true}{here}
}
