% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set_global_opts.R
\name{get_style_path}
\alias{get_style_path}
\title{Get style sheet path}
\usage{
get_style_path()
}
\description{
Returns path to current style sheet which is set in global options. To change this path see \link\link{xltabr}{set_style_path}.
Default stylesheet used by package is \href{https://github.com/moj-analytical-services/xltabr/blob/dev/inst/extdata/styles.xlsx?raw=true}{here}
}
