% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{\%>\%#'}
\alias{\%>\%#'}
\alias{initialise}
\title{Create a new xltabr object for cross tabulation}
\usage{
initialise(wb = NULL, ws_name = NULL, styles_xlsx = NULL,
  num_styles_csv = NULL, topleft_row = 1, topleft_col = 1,
  insert_below_tab = NULL)
}
\arguments{
\item{insert_below_tab}{If given, the new tab will be inserted immediately below this one}

\item{df}{The dataframe you want to output to Excel}
}
\value{
A list which contains the dataframe
}
\description{
Create a new xltabr object for cross tabulation
}
