% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{initialise}
\alias{initialise}
\title{Create a new xltabr object for cross tabulation}
\usage{
initialise(wb = NULL, ws_name = NULL, topleft_row = 1, topleft_col = 1,
  insert_below_tab = NULL)
}
\arguments{
\item{wb}{An openxlsx workbook object to write to.  If null, a new one will be created}

\item{ws_name}{The worksheet to write to.  If null, Sheet1 will be used}

\item{topleft_row}{Specifies the row where the table begins in the worksheet}

\item{topleft_col}{Specifies the column where the table begins in the worksheet}

\item{insert_below_tab}{If given, the new tab will be inserted immediately below this one}
}
\value{
A list which contains the dataframe
}
\description{
Create a new xltabr object for cross tabulation
}
\examples{
tab <- initialise()
}
