% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set_global_opts.R
\name{set_cell_format_path}
\alias{set_cell_format_path}
\title{Set cell format path}
\usage{
set_cell_format_path(path = NULL)
}
\arguments{
\item{path}{the file path to the cell formats (csv file). If NULL the function sets the cell format to the default option.}
}
\description{
Set the path to the cell formats to be used by xltabr. To get this path see get_cell_format_path.
Default cell format used by package is \href{https://github.com/moj-analytical-services/xltabr/blob/master/inst/extdata/style_to_excel_number_format.csv?raw=true}{here}.
If no path is supplied the function sets the cell format to default to default.
}
\examples{
my_file_path <- system.file("extdata", "style_to_excel_number_format.csv", package = "xltabr")
set_cell_format_path(my_file_path)
}
