% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set_global_opts.R
\name{set_num_format_path}
\alias{set_num_format_path}
\title{Set number format path}
\usage{
set_num_format_path(path = NULL)
}
\arguments{
\item{path}{the file path to the number formats (csv file). If NULL the function sets the number format to the default option.}
}
\description{
Set the path to the number formats to be used by xltabr. To get this path see get_num_format_path.
Default number format used by package is \href{https://github.com/moj-analytical-services/xltabr/blob/dev/inst/extdata/number_format_defaults.csv?raw=true}{here}.
If no path is supplied the function sets the cell format to default to default.
}
