% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_catalogue.R
\name{style_catalogue_add_excel_num_format}
\alias{style_catalogue_add_excel_num_format}
\title{Manually add an Excel num format to the style catalogue}
\usage{
style_catalogue_add_excel_num_format(tab, style_string, excel_num_format)
}
\arguments{
\item{tab}{a table object}

\item{style_string}{the name (key) in the tab$style_catalogue}

\item{excel_num_format}{an excel number format e.g. "#.00"}
}
\description{
Manually add an Excel num format to the style catalogue
}
