% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_catalogue.R
\name{style_catalogue_override_styles}
\alias{style_catalogue_override_styles}
\title{Allows the user to provide a .xlsx with custom defined styles
Use \href{https://github.com/moj-analytical-services/xltabr/blob/dev/inst/extdata/styles.xlsx?raw=true}{this} file as a basis}
\usage{
style_catalogue_override_styles(tab, path_to_xlsx)
}
\arguments{
\item{path}{a file path to the .xlsx file you want to use to override styles}
}
\value{
The tab
}
\description{
Allows the user to provide a .xlsx with custom defined styles
Use \href{https://github.com/moj-analytical-services/xltabr/blob/dev/inst/extdata/styles.xlsx?raw=true}{this} file as a basis
}
