% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{title_get_wb_cols}
\alias{title_get_wb_cols}
\title{Get the columns occupied by the title in the workbook}
\usage{
title_get_wb_cols(tab)
}
\arguments{
\item{tab}{The core tab object}
}
\description{
Get the columns occupied by the title in the workbook
}
