% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{title_initialise}
\alias{title_initialise}
\title{Create all the required properties for the title on the tab object}
\usage{
title_initialise(tab)
}
\arguments{
\item{tab}{core tab object}
}
\description{
Create all the required properties for the title on the tab object
}
