% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_headers.R
\name{top_headers_get_bottom_wb_row}
\alias{top_headers_get_bottom_wb_row}
\title{Compute the bottom (lowest) row of the workbook occupied by the top headers
If the top headers do not exist, returns the last row of the previous element (the titles)}
\usage{
top_headers_get_bottom_wb_row(tab)
}
\arguments{
\item{tab}{The core tab object}
}
\description{
Compute the bottom (lowest) row of the workbook occupied by the top headers
If the top headers do not exist, returns the last row of the previous element (the titles)
}
