% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_headers.R
\name{top_headers_get_rightmost_wb_col}
\alias{top_headers_get_rightmost_wb_col}
\title{Compute the rightmost column of the workbook which is occupied by the top headers}
\usage{
top_headers_get_rightmost_wb_col(tab)
}
\arguments{
\item{tab}{The core tab object}
}
\description{
Compute the rightmost column of the workbook which is occupied by the top headers
}
