% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_headers.R
\name{top_headers_get_wb_cols}
\alias{top_headers_get_wb_cols}
\title{Compute the columns of the workbook which are occupied by the top headers}
\usage{
top_headers_get_wb_cols(tab)
}
\arguments{
\item{tab}{The core tab object}
}
\description{
Compute the columns of the workbook which are occupied by the top headers
}
