% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_headers.R
\name{top_headers_initialise}
\alias{top_headers_initialise}
\title{Create all the required properties for the top headers on the tab object}
\usage{
top_headers_initialise(tab)
}
\arguments{
\item{tab}{The core tab object}
}
\description{
Create all the required properties for the top headers on the tab object
}
