% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_headers.R
\name{top_headers_write_rows}
\alias{top_headers_write_rows}
\title{Write all the required data (but no styles)}
\usage{
top_headers_write_rows(tab)
}
\description{
Write all the required data (but no styles)
}
