% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer.R
\name{add_footer}
\alias{add_footer}
\title{Add footers to the tab.  Footer text is provided as a character vector, with each element being a row of the footer}
\usage{
add_footer(tab, footer_text, footer_style_names = "footer")
}
\arguments{
\item{footer_text}{A character vector.  Each element is a row of the footer}

\item{footer_style_names}{A character vector.  Each elemment is a style_name}
}
\description{
Add footers to the tab.  Footer text is provided as a character vector, with each element being a row of the footer
}
