% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{auto_merge_footer_cells}
\alias{auto_merge_footer_cells}
\title{Take a tab and merge the footer rows}
\usage{
auto_merge_footer_cells(tab)
}
\arguments{
\item{df}{A tab}
}
\description{
Take a tab and merge the footer rows
}
