% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{auto_merge_title_cells}
\alias{auto_merge_title_cells}
\title{Take a tab and merge the title rows}
\usage{
auto_merge_title_cells(tab)
}
\arguments{
\item{tab}{The core tab object}
}
\description{
Take a tab and merge the title rows
}
