% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_style.R
\name{auto_style_number_formatting}
\alias{auto_style_number_formatting}
\title{Use the data type of the columns to choose an automatic Excel format for body cells}
\usage{
auto_style_number_formatting(tab, overrides = list())
}
\arguments{
\item{tab}{a table object}

\item{overrides}{a list containing any manual overrides where the user wants to provide their own style name}
}
\description{
This function reads currency styling from the styles defined \href{https://github.com/moj-analytical-services/xltabr/blob/master/inst/extdata/number_formats.xlsx?raw=true}{here}
}
