% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{body_get_wb_left_header_cols}
\alias{body_get_wb_left_header_cols}
\title{Get the subset of body columns which are left header columns}
\usage{
body_get_wb_left_header_cols(tab)
}
\description{
Get the subset of body columns which are left header columns
}
