% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{body_initialise}
\alias{body_initialise}
\title{Initialise the body, creating all the required properties the core tab object}
\usage{
body_initialise(tab)
}
\arguments{
\item{tab}{The core tab object}
}
\description{
Initialise the body, creating all the required properties the core tab object
}
