% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{title_get_bottom_wb_row}
\alias{title_get_bottom_wb_row}
\title{Get the bottom row which the titles occupy in the workbook
If no titles have been provided, return the cell above the topleft row of the extent.
The next element (top headers, body or whatever it is) will want to be positioned
in the cell below this}
\usage{
title_get_bottom_wb_row(tab)
}
\arguments{
\item{tab}{The core tab object}
}
\description{
Get the bottom row which the titles occupy in the workbook
If no titles have been provided, return the cell above the topleft row of the extent.
The next element (top headers, body or whatever it is) will want to be positioned
in the cell below this
}
