% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{title_get_cell_styles_table}
\alias{title_get_cell_styles_table}
\title{Create table |row|col|style name| containing the styles names}
\usage{
title_get_cell_styles_table(tab)
}
\description{
Create table |row|col|style name| containing the styles names
}
