% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{title_get_rightmost_wb_col}
\alias{title_get_rightmost_wb_col}
\title{Get the rightmost column of the titles in the wb}
\usage{
title_get_rightmost_wb_col(tab)
}
\description{
Get the rightmost column of the titles in the wb
}
