% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_headers.R
\name{top_headers_get_cell_styles_table}
\alias{top_headers_get_cell_styles_table}
\title{Create table |row|col|style name| containing the styles names}
\usage{
top_headers_get_cell_styles_table(tab)
}
\description{
Create table |row|col|style name| containing the styles names
}
